requirements.txt
django==4.0.2
django-environ==0.4.5
psycopg2-binary==2.8.6
djangorestframework==3.13.1